backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.28.85
python-dotenv==1.0.0
PyMuPDF==1.23.8
Pillow==10.0.1
easyocr==1.7.0
openpyxl==3.1.2
requests==2.31.0